+++ src/structures/paging/mapper/mod.rs
@@ -1,8 +1,8 @@
 //! Abstractions for reading and modifying the mapping of pages.
 
-pub use self::mapped_page_table::MappedPageTable;
+pub use self::mapped_page_table::{MappedPageTable, PhysToVirt};
 #[cfg(target_arch = "x86_64")]
-pub use self::recursive_page_table::RecursivePageTable;
+pub use self::{offset_page_table::OffsetPageTable, recursive_page_table::RecursivePageTable};
 
 use crate::structures::paging::{
     frame_alloc::FrameAllocator, page_table::PageTableFlags, Page, PageSize, PhysFrame, Size1GiB,
@@ -11,6 +11,7 @@ use crate::structures::paging::{
 use crate::{PhysAddr, VirtAddr};
 
 mod mapped_page_table;
+mod offset_page_table;
 mod recursive_page_table;
 
 /// This trait defines page table operations that work for all page sizes of the x86_64